requirements.txt
matplotlib==3.9.2
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
Pillow>=8,<11
streamlit==1.38.0
